+++ functions/wp_bootstrap_navwalker.php
@@ -82,9 +82,17 @@ class wp_bootstrap_navwalker extends Walker_Nav_Menu {
 
 			// If item has_children add atts to a.
 			if ( $args->has_children && $depth === 0 ) {
-				$atts['href']   		= '#';
-				$atts['data-toggle']	= 'dropdown';
-				$atts['class']			= 'dropdown-toggle';
+				
+				if(ot_get_option('gp-nav-hover-dropdowns', 'off') == 'on')
+				{
+					$atts['href'] = ! empty( $item->url ) ? $item->url : '';
+				}
+				else 
+				{
+					$atts['href']   		= '#';
+					$atts['data-toggle']	= 'dropdown';
+					$atts['class']			= 'dropdown-toggle';
+				}
 				$atts['aria-haspopup']	= 'true';
 			} else {
 				$atts['href'] = ! empty( $item->url ) ? $item->url : '';